skin_disease_prediction/requirements.txt
Flask==2.0.1
tensorflow==2.8.0
numpy==1.22.3
Pillow==9.0.1
Werkzeug==2.0.2
gunicorn==20.1.0